+++ CHANGE_LOG.md
@@ -9,3 +9,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Ability to analyze built source for unreferenced classes
 - Added annotation usage of classes to dead class analysis
 - Added ability to filter classes analyzed
+- Added implemented interfaces to dead class analysis
+- Added thrown exceptions to dead class analysis